+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Support for web origin-bound responses, if the client has a registered web
+origin and specifies that origin as the value of `redirect_uri` in the
+authorization request.
 - Support for redirect URI schemes, enabling support for [IDP-IFrame-based Implict Flow](http://lists.openid.net/pipermail/openid-specs-ab/Week-of-Mon-20151116/005865.html)
 and the storagerelay scheme.